🏷️ move search to api, artist to resource
